fiddling around with npm import methodologies
