+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sh-core",
-  "version": "0.1.9",
+  "version": "0.1.10",
   "description": "Base SASS and React files for SuperheroUI components",
   "main": "bin/sh-core.js",
   "scripts": {
