--- src/util/index.js
@@ -1,7 +1,0 @@
-import ClassService from './class.service';
-
-export {ClassService};
-
-export default {
-    ClassService
-}
